refactor mainlayout component and update locale moment
