+++ src/modules/inbox/components/RespondBox.js
@@ -205,7 +205,7 @@ class RespondBox extends Component {
       return (
         <Mask onClick={this.hideMask}>
           {__(
-            'Customer is offline. Click to hide and send messages and they will receive them the next time they are online.' // eslint-disable-line
+            'Customer is offline. Click to hide and send messages and they will receive them the next time they are online.'
           )}
         </Mask>
       );
