+++ src/modules/layout/components/MainLayout.js
@@ -1,99 +1,20 @@
 import React from 'react';
-import moment from 'moment';
 import { withRouter } from 'react-router';
 import PropTypes from 'prop-types';
 import { Layout } from '../styles';
 import { Navigation } from '../containers';
-import translations from '../../../locales';
-import en from 'react-intl/locale-data/en';
-import mn from 'react-intl/locale-data/mn';
-import { injectIntl, IntlProvider, addLocaleData } from 'react-intl';
 
 const propTypes = {
   history: PropTypes.object,
   currentUser: PropTypes.object,
-  selectLang: PropTypes.func,
-  locale: PropTypes.string,
   children: PropTypes.node
 };
 
-class TranslationWrapper extends React.Component {
-  getChildContext() {
-    const { intl } = this.props;
-    const { formatMessage } = intl;
-
-    return {
-      __: msg => formatMessage({ id: msg })
-    };
-  }
-
-  render() {
-    const { children, currentUser } = this.props;
-    return (
-      <React.Fragment>
-        {currentUser && <Navigation />}
-        {children}
-      </React.Fragment>
-    );
-  }
-}
-
-TranslationWrapper.propTypes = {
-  intl: PropTypes.object,
-  children: PropTypes.object,
-  currentUser: PropTypes.object
-};
-
-TranslationWrapper.childContextTypes = {
-  __: PropTypes.func
-};
-
-const InjectedComponent = injectIntl(TranslationWrapper);
-
-const messages = {
-  ...translations
-};
-
-addLocaleData([...en, ...mn]);
-
 class MainLayout extends React.Component {
-  constructor(props) {
-    super(props);
-
-    this.state = {
-      locale: 'en',
-      messages: messages['en']
-    };
-
-    this.selectLang = this.selectLang.bind(this);
-  }
-
   getChildContext() {
     return {
-      currentUser: this.props.currentUser,
-      selectLang: this.selectLang,
-      locale: this.state.locale
+      currentUser: this.props.currentUser
     };
-  }
-
-  componentWillMount() {
-    moment.locale('mn', {
-      relativeTime: {
-        future: '%s дараа',
-        past: '%s өмнө',
-        s: 'саяхан',
-        m: 'минутын',
-        mm: '%d минутын',
-        h: '1 цагийн',
-        hh: '%d цагийн',
-        d: '1 өдрийн',
-        dd: '%d өдрийн',
-        M: '1 сарын',
-        MM: '%d сарын',
-        y: '1 жилийн',
-        yy: '%d жилийн'
-      }
-    });
   }
 
   componentDidMount() {
@@ -116,36 +37,16 @@ class MainLayout extends React.Component {
       })(),
       true
     );
-
-    this.getLang();
-  }
-
-  getLang() {
-    const locale = localStorage.getItem('locale');
-
-    this.selectLang(locale);
-  }
-
-  selectLang(locale) {
-    localStorage.setItem('locale', locale || 'en');
-
-    moment.locale(locale || 'en');
-
-    this.setState({
-      locale: locale || 'en',
-      messages: messages[locale || 'en']
-    });
   }
 
   render() {
-    const { locale, messages } = this.state;
+    const { currentUser, children } = this.props;
 
     return (
-      <IntlProvider locale={locale} messages={messages}>
-        <Layout>
-          <InjectedComponent {...this.props} />
-        </Layout>
-      </IntlProvider>
+      <Layout>
+        {currentUser && <Navigation />}
+        {children}
+      </Layout>
     );
   }
 }
@@ -153,9 +54,7 @@ class MainLayout extends React.Component {
 MainLayout.propTypes = propTypes;
 
 MainLayout.childContextTypes = {
-  currentUser: PropTypes.object,
-  selectLang: PropTypes.func,
-  locale: PropTypes.string
+  currentUser: PropTypes.object
 };
 
 export default withRouter(MainLayout);
